move makeConfig into its own file
