+++ node/lib/client.js
@@ -5,13 +5,13 @@
 
 var Promise = require('bluebird');
 var ArgumentError = require('azure-iot-common').errors.ArgumentError;
-var ConnectionString = require('azure-iot-common').ConnectionString;
+var makeConfig = require('./config.js');
 
 function EventHubClient(config) {
   if (!config.host) throw new ArgumentError('Argument config is missing property host');
 }
 
-EventHubClient.fromConnectionString = function (connectionString, path, Type_) {
+EventHubClient.fromConnectionString = function (connectionString, path) {
   if (!connectionString) {
     throw new ArgumentError('Missing argument connectionString');
   }
@@ -21,23 +21,11 @@ EventHubClient.fromConnectionString = function (connectionString, path, Type_) {
     throw new ArgumentError('Connection string doesn\'t have EntityPath, or missing argument path');
   }
 
-  Type_ = Type_ || EventHubClient;
-
-  return new Type_(config);
+  return new EventHubClient(config);
 };
 
 EventHubClient.prototype.open = function () {
   return Promise.resolve();
 };
-
-function makeConfig(connectionString, path) {
-  var cn = ConnectionString.parse(connectionString);
-  var endpoint = cn.Endpoint || ''; 
-  
-  return {
-    host: (endpoint.match('sb://([^/]*)') || [])[1],
-    path: cn.EntityPath || path
-  };
-}
 
 module.exports = EventHubClient;
