+++ node/tests/client_creation.js
@@ -6,6 +6,7 @@
 var chai = require('chai');
 chai.should();
 
+var makeConfig = require('../lib/config.js');
 var EventHubClient = require('../lib/client.js');
 var ArgumentError = require('azure-iot-common').errors.ArgumentError;
 
@@ -14,6 +15,14 @@ function testFalsyValues(testFn) {
     testFn(value);
   });
 }
+
+describe('makeConfig', function () {
+  it('populates config.host from the connection string\'s Endpoint', function () {
+    var config = makeConfig('Endpoint=sb://abc');
+    config.should.have.property('host')
+      .that.equals('abc');
+  });
+});
 
 describe('EventHubClient', function () {
   describe('#constructor', function () {
@@ -44,13 +53,6 @@ describe('EventHubClient', function () {
       test.should.throw(ArgumentError, 'Connection string doesn\'t have EntityPath, or missing argument path');
     });
     
-    it('populates config.host from the connection string\'s Endpoint', function () {
-      function TestClient(config) { this.config = config; }
-      var client = EventHubClient.fromConnectionString('Endpoint=sb://abc;EntityPath=xyz', null, TestClient);
-      client.config.should.have.property('host')
-        .that.equals('abc');
-    });
-
     it('creates an EventHubClient from a connection string', function () {
       var client = EventHubClient.fromConnectionString('Endpoint=sb://abc;EntityPath=xyz');
       client.should.be.an.instanceof(EventHubClient);
